+++ src/realm/history.cpp
@@ -27,7 +27,7 @@ using namespace realm;
 
 namespace {
 
-// As new schema versions come into existsnece, describe them here.
+// As new schema versions come into existence, describe them here.
 constexpr int g_history_schema_version = 0;
 
 
@@ -212,7 +212,7 @@ void InRealmHistory::update_from_ref(ref_type ref, version_type version)
 }
 
 
-class InRealmHistoryImpl : public TrivialReplication, private InRealmHistory {
+class InRealmHistoryImpl : public TrivialReplication {
 public:
     using version_type = TrivialReplication::version_type;
 
@@ -225,7 +225,7 @@ public:
     {
         TrivialReplication::initialize(sg); // Throws
         using sgf = _impl::SharedGroupFriend;
-        InRealmHistory::initialize(sgf::get_group(sg)); // Throws
+        m_history.initialize(sgf::get_group(sg)); // Throws
     }
 
     void initiate_session(version_type) override
@@ -241,9 +241,9 @@ public:
     version_type prepare_changeset(const char* data, size_t size, version_type orig_version) override
     {
         if (!is_history_updated())
-            update_from_parent(orig_version); // Throws
+            m_history.update_from_parent(orig_version); // Throws
         BinaryData changeset(data, size);
-        version_type new_version = add_changeset(changeset); // Throws
+        version_type new_version = m_history.add_changeset(changeset); // Throws
         return new_version;
     }
 
@@ -280,13 +280,11 @@ public:
 
     _impl::History* get_history() override
     {
-        return this;
+        return &m_history;
     }
 
-    BinaryData get_uncommitted_changes() noexcept override
-    {
-        return TrivialReplication::get_uncommitted_changes();
-    }
+private:
+    InRealmHistory m_history;
 };
 
 } // unnamed namespace