+++ src/realm/replication.hpp
@@ -180,6 +180,17 @@ public:
 
     //@}
 
+    /// Get the list of uncommited changes accumulated so far in the current
+    /// write transaction.
+    ///
+    /// The callee retains ownership of the referenced memory. The ownership is
+    /// not handed over the the caller.
+    ///
+    /// This function may be called only during a write transaction (prior to
+    /// initiation of commit operation). In that case, the caller may assume that the
+    /// returned memory reference stays valid for the remainder of the transaction (up
+    /// until initiation of the commit operation).
+    virtual BinaryData get_uncommitted_changes() const noexcept = 0;
 
     /// Interrupt any blocking call to a function in this class. This function
     /// may be called asyncronously from any thread, but it may not be called
@@ -408,7 +419,7 @@ protected:
 
     bool is_history_updated() const noexcept;
 
-    BinaryData get_uncommitted_changes() const noexcept;
+    BinaryData get_uncommitted_changes() const noexcept override;
 
     std::string get_database_path() override;
     void initialize(DB&) override;